public variables with capital case
