--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A1.A.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f5531b68f10be264baebf8e7bfde3ff4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
