--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A1.A/Q.B1.asset
@@ -1,25 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 27f5ce73c4fbedd4ebf32b7dc1f44bb2, type: 3}
-  m_Name: Q.B1
-  m_EditorClassIdentifier: 
-  Text: Alright then. Well, if you must know, I was in the rooftop garden, pruning
-    the roses. Mr Jones can vouch for me.
-  choices:
-  - Text: Pruning the roses...at night?
-    dialogue: {fileID: 0}
-    question: {fileID: 0}
-  - Text: 'That checks out. You weren''t near the crime scene at all.
-
-'
-    dialogue: {fileID: 0}
-    question: {fileID: 0}
