--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A1.A/Q.B1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9ae19dd19c1793846a4b1eb8c947ac43
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
