--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A1.A/Q.B2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e47e7edd8bd7a3745b3875bbf471f05c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
