--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A1.D.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6f58d2f2c4326554192b1344512242ec, type: 3}
-  m_Name: A1
-  m_EditorClassIdentifier: 
-  sentences:
-  - Text: Last Tuesday, let me think... oh, that must have been the horrid day that
-      Julia passed away. Such an awful thing to happen. Sweet girl, she was.
-  question: {fileID: 11400000, guid: 3f2bac8682d0d7e4bbbdbd3519343bb3, type: 2}
