--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A1.D.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aecf4e6ff1debbd46a530caf0f0eb78a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
