--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A1.Q.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3f2bac8682d0d7e4bbbdbd3519343bb3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
