--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A2.A.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0786da7bd780ddf4cb3d4d89bb2c75dc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
