--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A2.Q.asset
@@ -1,24 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 27f5ce73c4fbedd4ebf32b7dc1f44bb2, type: 3}
-  m_Name: A2
-  m_EditorClassIdentifier: 
-  Text: Oh... um, I'd rather not talk about it. It's, uh, a painful topic.
-  choices:
-  - Text: We can talk about something else?
-    dialogue: {fileID: 0}
-    question: {fileID: 0}
-  - Text: 'I really need to know.
-
-'
-    dialogue: {fileID: 0}
-    question: {fileID: 0}
