--- Assets/Resources/ScriptableObjects/Conversations/Policeman/Answers/A2.Q.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6d256f4072b7f854a8f65c4c17769e81
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
