+++ Assets/Resources/ScriptableObjects/Conversations/Policeman/New Question.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: f5006f5d804f9544e8f1734bba70a21b
-folderAsset: yes
-DefaultImporter:
+guid: fe3d0fde11418dc4996f7c2a4a0008db
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
