--- Assets/Resources/ScriptableObjects/Conversations/Policeman/OpeningQuestion.asset
@@ -1,28 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 27f5ce73c4fbedd4ebf32b7dc1f44bb2, type: 3}
-  m_Name: OpeningQuestion
-  m_EditorClassIdentifier: 
-  Text: Tell me, what do you want to talk about?
-  choices:
-  - ProffesionUnlocksChoice: 1
-    Text: What were you doing last Tuesday night?
-    dialogue: {fileID: 11400000, guid: aecf4e6ff1debbd46a530caf0f0eb78a, type: 2}
-    question: {fileID: 0}
-  - ProffesionUnlocksChoice: 0
-    Text: Did you ever find out what happened to your son?
-    dialogue: {fileID: 0}
-    question: {fileID: 11400000, guid: 6d256f4072b7f854a8f65c4c17769e81, type: 2}
-  - ProffesionUnlocksChoice: 0
-    Text: Nothing, at the moment.
-    dialogue: {fileID: 0}
-    question: {fileID: 0}
