--- Assets/Resources/ScriptableObjects/Conversations/Policeman/OpeningQuestion.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 82ed04bda7c273048acd99daa122f3a4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
