+++ Assets/Resources/ScriptableObjects/Dialogue/Choice.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e29407fa940a1244b964ed0033d7cc50
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
