+++ Assets/Resources/ScriptableObjects/Dialogue/Question.cs
@@ -1,20 +1,9 @@
-﻿using Interfaces;
-using UnityEngine;
-
-[System.Serializable]
-public struct Choice
-{
-    public Proffesion ProffesionUnlocksChoice;
-    [TextArea(2, 5)]
-    public string Text;
-    public Dialogue dialogue;
-    public Question question;
-}
+﻿using UnityEngine;
 
 [CreateAssetMenu(fileName = "New Question", menuName = "Question")]
 public class Question : ScriptableObject
 {
     [TextArea(2, 5)]
     public string Text;
-    public Choice[] choices;
+    public Choice[] Choices;
 }
